+++ BAVCL/Core/GPU/Memory Management/Cache.cs
